Creation of ``qa-util/common`` module, and changing all uses of ``ql.type.DateUtils`` and ``ql.util.DateUtils`` use it instead
