+++ x-pack/plugin/eql/build.gradle
@@ -20,6 +20,8 @@ dependencies {
   compileOnly(project(':modules:lang-painless:spi'))
   compileOnly project(xpackModule('ql'))
 
+  compileOnly project(xpackModule('qa-util:common'))
+
   testImplementation project(':test:framework')
   testImplementation(testArtifact(project(xpackModule('core'))))
   testImplementation(testArtifact(project(xpackModule('security'))))
