+++ x-pack/plugin/eql/src/test/java/org/elasticsearch/xpack/eql/execution/sequence/PITFailureTests.java
@@ -69,7 +69,7 @@ public class PITFailureTests extends ESTestCase {
 
             EqlConfiguration eqlConfiguration = new EqlConfiguration(
                 new String[] { "test" },
-                org.elasticsearch.xpack.ql.util.DateUtils.UTC,
+                org.elasticsearch.xpack.qautil.DateUtils.UTC,
                 "nobody",
                 "cluster",
                 null,
