+++ x-pack/plugin/qa-util/build.gradle
@@ -1,0 +1,5 @@
+apply plugin: 'elasticsearch.java'
+
+dependencies {
+  api project(':test:framework')
+}
