+++ x-pack/plugin/ql/build.gradle
@@ -18,6 +18,9 @@ dependencies {
   api "org.antlr:antlr4-runtime:${antlrVersion}"
   api project(path: xpackModule('mapper-version'))
   compileOnly project(path: xpackModule('core'))
+
+  compileOnly project(xpackModule('qa-util:common'))
+
   testApi(project(xpackModule('ql:test-fixtures'))) {
     exclude group: 'org.elasticsearch.plugin', module: 'ql'
   }
