+++ x-pack/plugin/ql/src/test/java/org/elasticsearch/xpack/ql/expression/predicate/RangeTests.java
@@ -10,7 +10,7 @@ package org.elasticsearch.xpack.ql.expression.predicate;
 import org.elasticsearch.test.ESTestCase;
 import org.elasticsearch.xpack.ql.tree.Source;
 import org.elasticsearch.xpack.ql.type.DataType;
-import org.elasticsearch.xpack.ql.type.DateUtils;
+import org.elasticsearch.xpack.qautil.DateUtils;
 
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
